Add skills and battlefield

+++ DDEngine/DDEngine/data/items/Item.cs
@@ -10,7 +10,7 @@ namespace DDEngine.data.items
     {
         public string Name { set; get; }
         public string Description { set; get; }
-        public int Weight { set; get; }
+        public float Weight { set; get; }
         public int Price { set; get; }
 
     }
